+++ rearq/server/app.py
@@ -5,10 +5,20 @@ from starlette.requests import Request
 from starlette.responses import JSONResponse
 from starlette.staticfiles import StaticFiles
 
-from .. import constants
+from .. import ReArq, constants
 from .routes import router
 
-app = FastAPI(title="API docs of ReArq")
+
+class App(FastAPI):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.rearq = None
+
+    def set_rearq(self, rearq: ReArq):
+        self.rearq = rearq
+
+
+app = App(title="API docs of ReArq")
 app.mount(
     "/static",
     StaticFiles(directory=os.path.join(constants.STATIC_DIR, "rearq", "server", "static")),